% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common.R
\name{add_issue_meta}
\alias{add_issue_meta}
\title{Add issue metadata}
\usage{
add_issue_meta(
  df,
  issue,
  type = c("raw", "summary"),
  dataset = NULL,
  id_cols = c("SITEID", "SUBJECT_ID"),
  status = "New",
  first = Sys.Date(),
  last = Sys.Date()
)
}
\arguments{
\item{df_summary}{A summary data.frame}

\item{df_raw}{A raw data.frame}

\item{idx}{Optional index}

\item{label}{Optional label}
}
\description{
Add issue metadata
}
