% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ds.R
\name{check_ds}
\alias{check_ds}
\title{Disposition Dataset Check}
\usage{
check_ds(
  ds = ds6001,
  formeid = FORMEID,
  scr_dt = DSSTDAT2,
  scr_reas = DSDECOD_13,
  trt_dt = DSSTDAT3,
  trt_reas = DSDECOD_7,
  trtp_dt = DSSTDAT6,
  trtp_reas = DSDECOD_9,
  fup_dt = DSSTDAT9,
  fup_reas = DSDECOD_15,
  visit_info_df = visit_info_df
)
}
\arguments{
\item{ds}{Disposition dataset name}

\item{output_tab}{The name of the worksheet to create in the Excel output.}
}
\value{
A dataframe that specified each subject's disposition information.
}
\description{
This function checks the ds6001 dataset (for out-sourced study)
}
