% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common.R
\name{check_duplicate_visit_date}
\alias{check_duplicate_visit_date}
\title{Check for duplicate visits with the same date}
\usage{
check_duplicate_visit_date(
  datasets_pool,
  wb,
  target_vars = NULL,
  dataset_no = NULL,
  visit_info_df = NULL,
  output_tab = NULL,
  custom_code_list = NULL
)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{Workbook}

\item{target_vars}{Optional}

\item{dataset_no}{Datasets to exclude}

\item{visit_info_df}{Visit info}

\item{output_tab}{Output tab}

\item{custom_code_list}{Named list of custom check functions}
}
\description{
Check for duplicate visits with the same date
}
