% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common.R
\name{check_missing_key_vars}
\alias{check_missing_key_vars}
\title{Check for missing key variables}
\usage{
check_missing_key_vars(
  datasets_pool,
  wb,
  target_vars = NULL,
  dataset_no = NULL,
  visit_info_df = NULL
)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{An openxlsx Workbook}

\item{target_vars}{Key variables to check for some/all datasets}

\item{dataset_no}{Names of datasets to exclude}

\item{visit_info_df}{Visit info metadata}

\item{output_tab}{Optional output tab}
}
\value{
Writes results into workbook, invisibly returns NULL
}
\description{
Check for missing key variables
}
