% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common.R
\name{check_visit_after_ED}
\alias{check_visit_after_ED}
\title{Check for visits after ED}
\usage{
check_visit_after_ED(
  datasets_pool,
  wb,
  dataset_no = NULL,
  visit_info_df = NULL,
  output_tab = NULL,
  other_datasets = NULL
)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{Workbook}

\item{dataset_no}{Excluded datasets}

\item{visit_info_df}{Visit info metadata}

\item{output_tab}{Output tab}

\item{other_datasets}{Character vector of datasets to check}
}
\value{
A named list, each element is a data frame of records
with visits after ED for the corresponding dataset.
If no datasets have violations, returns an empty list.
}
\description{
Check for visits after ED
}
