% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_visit_gap.R
\name{check_visit_gap}
\alias{check_visit_gap}
\title{Check visit gaps between consecutive visits}
\usage{
check_visit_gap(
  datasets_pool,
  wb,
  visit_info_df = NULL,
  output_tab = NULL,
  day_lob = 7,
  day_hib = 21
)
}
\arguments{
\item{datasets_pool}{Named list of raw datasets}

\item{wb}{An openxlsx Workbook object}

\item{visit_info_df}{Metadata with dataset, visit_label_col, visit_date_col}

\item{output_tab}{Name of Excel worksheet to create}

\item{day_lob}{Minimum day gap (default: 7)}

\item{day_hib}{Maximum day gap (default: 21)}
}
\value{
Writes a worksheet with detected issues into \code{wb}
}
\description{
This function checks that consecutive scheduled visits are more than
\code{day_lob} days (default: 7 days / 1 week) and less than
\code{day_hib} days (default: 21 days / 3 weeks) apart.
}
\details{
Visits such as 1, 801, 802, 803 are excluded from the check.
}
