% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress_ranges}
\alias{compress_ranges}
\title{Compress consecutive integers into ranges}
\usage{
compress_ranges(x)
}
\arguments{
\item{x}{Integer vector}
}
\value{
A string such as "1-4, 6"
}
\description{
Compress consecutive integers into ranges
}
