% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dup.R
\name{compute_duplicate_visit_date}
\alias{compute_duplicate_visit_date}
\title{Check for duplicate visits with the same date}
\usage{
compute_duplicate_visit_date(df, ds, visit_date_col, visit_label_col, id_cols)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{Workbook}

\item{target_vars}{Optional}

\item{dataset_no}{Datasets to exclude}

\item{visit_info_df}{Visit info}

\item{output_tab}{Output tab}

\item{custom_code_list}{Named list of custom check functions}
}
\description{
Check for duplicate visits with the same date
}
