% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_misskeys.R
\name{compute_missing_key_vars}
\alias{compute_missing_key_vars}
\title{Check for missing key variables}
\usage{
compute_missing_key_vars(df, ds, keys)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{An openxlsx Workbook}

\item{target_vars}{Key variables to check for some/all datasets}

\item{dataset_no}{Names of datasets to exclude}

\item{visit_info_df}{Visit info metadata}
}
\value{
Writes results into workbook, invisibly returns NULL
}
\description{
Check for missing key variables
}
