% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlap.R
\name{compute_visit_overlap}
\alias{compute_visit_overlap}
\title{Check for overlapped visits}
\usage{
compute_visit_overlap(
  df,
  ds,
  visit_date_col,
  visit_label_col,
  id_cols = c("SITEID", "SUBJECT_ID")
)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{Workbook}

\item{dataset_no}{Excluded datasets}

\item{visit_info_df}{Visit info metadata}

\item{output_tab}{Output tab}
}
\value{
A named list, each element is a data frame of records
with overlapped visits for the corresponding dataset.
If no datasets have overlaps, returns an empty list.
}
\description{
Check for overlapped visits
}
