% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dates_to_char}
\alias{convert_dates_to_char}
\title{Convert Date/POSIX columns in a data.frame to character}
\usage{
convert_dates_to_char(
  df,
  date_format = "\%Y-\%m-\%d",
  dttm_format = "\%Y-\%m-\%d \%H:\%M:\%S"
)
}
\arguments{
\item{df}{A data.frame}

\item{date_format}{Format for Date columns}

\item{dttm_format}{Format for POSIXct/POSIXlt columns}
}
\value{
data.frame with converted columns
}
\description{
Convert Date/POSIX columns in a data.frame to character
}
