% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_after_ED.R
\name{get_ed_dates}
\alias{get_ed_dates}
\title{Check for visits after ED}
\usage{
get_ed_dates(
  ed_ds,
  id_col = "SUBJECT_ID",
  date_col = "DSSTDAT",
  form_col = "FORMEID",
  forms = c("DS6001_LV6", "DS6001_LV5")
)
}
\arguments{
\item{datasets_pool}{Named list of datasets}

\item{wb}{Workbook}

\item{dataset_no}{Excluded datasets}

\item{visit_info_df}{Visit info metadata}

\item{output_tab}{Output tab}

\item{other_datasets}{Character vector of datasets to check}
}
\value{
A named list, each element is a data frame of records
with visits after ED for the corresponding dataset.
If no datasets have violations, returns an empty list.
}
\description{
Check for visits after ED
}
