% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_read.R
\name{load_dataset}
\alias{load_dataset}
\title{Load raw datasets into a named list, and also assign each to environment}
\usage{
load_dataset(
  path,
  pattern = "\\\\.sas7bdat$",
  exclude_prefixes = c("sys", "cdms"),
  exclude_names = c("record_count", "rtsm_sbjct_data"),
  envir = .GlobalEnv
)
}
\arguments{
\item{path}{Directory to search}

\item{pattern}{Regex for files (default .sas7bdat)}

\item{exclude_prefixes}{Character vector of prefixes to drop}

\item{exclude_names}{Exact dataset basenames to drop (without extension)}

\item{envir}{Environment to assign datasets into (default: .GlobalEnv)}
}
\value{
Named list of data.frames/tibbles
}
\description{
Load raw datasets into a named list, and also assign each to environment
}
