% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common.R
\name{raw_process}
\alias{raw_process}
\title{Post-process raw issue records}
\usage{
raw_process(raw, issue)
}
\arguments{
\item{raw}{A data.frame of issues}

\item{issue}{A string with issue description}
}
\value{
data.frame with added metadata columns
}
\description{
Post-process raw issue records
}
\keyword{internal}
