% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchk0-package.R
\docType{package}
\name{rchk0-package}
\alias{rchk0}
\alias{rchk0-package}
\title{rchk0: Raw Data Issue Tracker for Clinical Datasets}
\description{
Tools for checking raw clinical datasets, identifying issues, and generating formatted Excel reports. Includes automated checks for visit discrepancies, early discontinuation dates, visit gaps, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Geomisgytree/rchk0}
  \item Report bugs at \url{https://github.com/Geomisgytree/rchk0/issues}
}

}
\author{
\strong{Maintainer}: Shushun Ren \email{shushunr@umich.edu}

}
\keyword{internal}
