% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tracker.R
\name{run_tracker}
\alias{run_tracker}
\title{Run Raw Issue Tracker at one time}
\usage{
run_tracker(
  data_path,
  spec_path,
  out_xlsx,
  study_name = NULL,
  exclude_prefixes = c("cdms"),
  exclude_names = c("record_count", "rtsm_sbjct_data", "sd", "mhpresp1001")
)
}
\arguments{
\item{data_path}{directory of raw datasets}

\item{spec_path}{xlsx spec path}

\item{out_xlsx}{output Excel path}

\item{study_name}{optional study label}

\item{exclude_prefixes}{prefixes to drop}

\item{exclude_names}{exact basenames to drop}
}
\description{
Run Raw Issue Tracker at one time
}
