% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_as_character}
\alias{safe_as_character}
\title{Safely convert to character while preserving NA}
\usage{
safe_as_character(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Character vector with NA unchanged
}
\description{
Safely convert to character while preserving NA
}
\keyword{internal}
