% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ds.R
\name{check_ds}
\alias{check_ds}
\title{Disposition Dataset Check}
\usage{
check_ds(
  ds = ds6001,
  formeid = FORMEID,
  scr_dt = NULL,
  scr_reas = NULL,
  trt_dt = NULL,
  trt_reas = NULL,
  trtp_dt = NULL,
  trtp_reas = NULL,
  fup_dt = NULL,
  fup_reas = NULL,
  visit_info_df = visit_info_df
)
}
\arguments{
\item{ds}{Disposition dataset name}

\item{output_tab}{The name of the worksheet to create in the Excel output.}
}
\value{
A dataframe that specified each subject's disposition information.
}
\description{
This function checks the DS6001 dataset (for out-sourced study)
}
