% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ds.R
\name{check_ds}
\alias{check_ds}
\title{Disposition Dataset Check}
\usage{
check_ds(ds, formeid, dataset_no = NULL, visit_info_df = NULL)
}
\arguments{
\item{ds}{Disposition dataset name}

\item{dataset_no}{A character vector of dataset names to exclude (e.g., diary data).}

\item{visit_info_df}{A data frame describing dataset, subject, site, visit/date column info. Please set}

\item{output_tab}{The name of the worksheet to create in the Excel output.}
}
\value{
The workbook \code{wb} with a new sheet added if issues are found.
}
\description{
This function performs two checks:
\enumerate{
\item For subjects having both ED and V801, ED date is on or after V801.
\item Regulatory visits (V1–V10) occurring after ED across all datasets (except diary data).
}
}
