% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_multi_site.R
\name{check_multi_site}
\alias{check_multi_site}
\title{Check subjects belonging to multiple sites}
\usage{
check_multi_site(datasets_pool, wb, output_tab = NULL, visit_info_df = NULL)
}
\arguments{
\item{datasets_pool}{Named list of raw datasets}

\item{wb}{An openxlsx Workbook object}

\item{output_tab}{Name of Excel worksheet to create}

\item{visit_info_df}{Metadata with dataset, subject_var, site_var}
}
\value{
Writes a worksheet with detected issues into \code{wb}
}
\description{
This function identifies subjects who appear in more than one site across datasets.
}
\keyword{internal}
