% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess raw datasets according to spec}
\usage{
preprocess_data(datasets_pool, preprocess_spec)
}
\arguments{
\item{datasets_pool}{Named list returned by load_dataset()}

\item{preprocess_spec}{Data frame from spec sheet PREPROCESSING}
}
\value{
list(visit_info_df=..., datasets_pool=...)
}
\description{
Preprocess raw datasets according to spec
}
