% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchk0-package.R
\docType{package}
\name{rchk0-package}
\alias{rchk0}
\alias{rchk0-package}
\title{rchk0: Raw Data Issue Tracker for Clinical Datasets}
\description{
Tools to detect, track, and export raw-data issues in clinical trials.
}
\details{
To get started, read the vignettes:
\code{browseVignettes(package = "rchk0")}
}
\section{Learn more}{

\itemize{
\item Vignettes: \code{browseVignettes("rchk0")}
}
}

\section{Useful links}{

\itemize{
\item Website: https://yourpkg-site.example
\item GitHub:  https://github.com/you/rchk0
\item Bug reports: https://github.com/you/rchk0/issues
}
}

\section{Authors}{

\itemize{
\item Maintainer: Shushun Ren \href{mailto:shushun.ren@lilly.com}{shushun.ren@lilly.com}
}
}

\seealso{
\link{rchk0-reference} for a categorized function index.
}
\author{
\strong{Maintainer}: Shushun Ren \email{shushunr@umich.edu}

}
\keyword{internal}
